Fix Bug 2466, prompt before taking over an existing bookmark's name.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@27352 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/text/fmt/xp/fv_View_cmd.cpp
@@ -5009,9 +5009,24 @@ UT_Error FV_View::cmdInsertBookmark(const char * szName)
 
 	if(!m_pDoc->isBookmarkUnique(static_cast<const gchar*>(szName)))
 	{
-		//bookmark already exists -- remove it and then reinsert
-		UT_DEBUGMSG(("fv_View::cmdInsertBookmark: bookmark \"%s\" exists - removing\n", szName));
-		_deleteBookmark(static_cast<const gchar*>(szName), false, &posStart, &posEnd);
+		XAP_Frame * pFrame = static_cast<XAP_Frame *>(getParentData());
+		XAP_Dialog_MessageBox::tAnswer ans = XAP_Dialog_MessageBox::a_NO;
+
+		if(pFrame)
+			ans = pFrame->showMessageBox(AP_STRING_ID_MSG_BookmarkAlreadyExists,
+									XAP_Dialog_MessageBox::b_YN, XAP_Dialog_MessageBox::a_NO);
+
+		if(ans == XAP_Dialog_MessageBox::a_YES)
+		{
+			//bookmark already exists -- remove it and then reinsert
+			UT_DEBUGMSG(("fv_View::cmdInsertBookmark: bookmark \"%s\" exists - removing\n", szName));
+			_deleteBookmark(static_cast<const gchar*>(szName), false, &posStart, &posEnd);
+		}
+		else
+		{
+			xxx_UT_DEBUGMSG(("User canceled bookmark replacement\n"));
+			return false;
+		}
 	}
 
 	gchar * pAttr[6];
